PythonScripts/requirements.txt
requests
azure-identity
pytesseract
pillow